requirements.txt
alembic==1.7.5
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==7.1.2
cryptography==36.0.1
firebase-admin==5.2.0
Flask==1.1.4
Flask-Cors==3.0.10
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
GeoAlchemy2==0.10.0
geographiclib==2.0
geopy==2.3.0
google-api-core==2.5.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.6
MarkupSafe==2.0.1
msgpack==1.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
packaging==21.3
pkg_resources==0.0.0
proto-plus==1.19.9
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.29
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==1.0.1
zipp==3.15.0
